pyproject.toml
[project]
name = "corapan_web"
version = "0.1.0"
description = "Next-generation CO.RA.PAN web application."
authors = [
  { name = "Felix Tacke", email = "felix.tacke@uni-marburg.de" }
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "Proprietary" }

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-dotenv>=0.5",
  "ruff>=0.5",
  "mypy>=1.10",
  "flask-testing>=0.8"
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"